feat : 방 아이디 요청 dto 통일(수동매칭 생성 dto는 분리)
