--- src/main/java/com/gachtaxi/domain/matching/common/dto/request/ManualMatchingJoinRequest.java
@@ -1,9 +1,0 @@
-package com.gachtaxi.domain.matching.common.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public record ManualMatchingJoinRequest(
-        @NotNull
-        Long roomId
-) {
-}
